tests/wpt/tests/content-security-policy/navigate-to/child-navigates-parent-blocked.sub.html

<!DOCTYPE html>

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  var t = async_test("Test that the child can't navigate the parent because the relevant policy belongs to the navigation initiator (in this case the child which has the policy `navigate-to 'none'`)");
  window.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.result, 'fail');
    assert_equals(e.data.violatedDirective, 'navigate-to');
  });
</script>
<iframe srcdoc="<iframe src='support/navigate_parent.sub.html?csp=navigate-to%20%27none%27&report_id={{$id:uuid()}}'>"></iframe>

<script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=navigate-to%20%27none%27&reportID={{$id}}'></script>
</body>
